Fix warnings

+++ src/utf8.h
@@ -1,7 +1,9 @@
 #ifndef _UTF8_H
 #define _UTF8_H
 
-extern unsigned long from_utf8(const unsigned char **inptr, int cnt, int *err);
+#include <stdint.h>
+
+extern int32_t from_utf8(const unsigned char **inptr, int cnt, int *err);
 
 #define UTF8_BAD (-1)
 #define UTF8_SHORT (-2)
